Adjust function signature to take quint16 instead of qint32

+++ src/knx/qknxnetipcrd.h
@@ -25,7 +25,7 @@ public:
     QKnxNetIpCRD() = default;
     ~QKnxNetIpCRD() override = default;
 
-    template <typename T> static QKnxNetIpCRD fromBytes(const T &bytes, qint32 index)
+    template <typename T> static QKnxNetIpCRD fromBytes(const T &bytes, quint16 index)
     {
         return QKnxNetIpStruct::fromBytes(bytes, index);
     }
