+++ src/knx/qknxnetipcurrentconfigdib.h
@@ -47,7 +47,7 @@ public:
                               const QHostAddress &dhcp,
                               AssignmentMethod method);
 
-    template <typename T> static QKnxNetIpCurrentConfigDIB fromBytes(const T &bytes, qint32 index)
+    template <typename T> static QKnxNetIpCurrentConfigDIB fromBytes(const T &bytes, quint16 index)
     {
         return QKnxNetIpStruct::fromBytes(bytes, index);
     }
