+++ src/knx/qknxnetipdevicedib.h
@@ -54,7 +54,7 @@ public:
                        const QByteArray &macAddress,
                        const QByteArray deviceName);
 
-    template <typename T> static QKnxNetIpDeviceDIB fromBytes(const T &bytes, qint32 index)
+    template <typename T> static QKnxNetIpDeviceDIB fromBytes(const T &bytes, quint16 index)
     {
         return QKnxNetIpStruct::fromBytes(bytes, index);
     }
