+++ tests/auto/qknxnetipstructure/tst_qknxnetipstructure.cpp
@@ -39,7 +39,7 @@ public:
     using QKnxNetIpStruct::toString;
 
     using QKnxNetIpStruct::bytes;
-    template<typename T> static TestStructure fromBytes(const T &bytes, qint32 index)
+    template<typename T> static TestStructure fromBytes(const T &bytes, quint16 index)
     {
         return QKnxNetIpStruct::fromBytes(bytes, index);
     }
